feat: Hallelujah thank You Jesus Christ our Holy Lord GOD Almighty

+++ Adding Ukrainian/Automate.py
@@ -11,6 +11,15 @@ Target_Path: str = os.path.join(
     "C:\\Program Files\\Epic Games\\CitiesSkylines\\Files\\Locale\\",
 )
 Current_Dir: str = os.path.dirname(os.path.abspath(__file__))
-shutil.copy2(
-    os.path.join(Current_Dir, "ua.locale"), os.path.join(Target_Path, "ua.locale")
-)
+with console.status("[yellow]Adding Ukrainian locale...[/yellow]"):
+    if os.path.exists(os.path.join(Target_Path, "ua.locale")):
+        console.print("Ukrainian locale already exists", style="bold cyan")
+        exit()
+    try:
+        shutil.copy2(
+            os.path.join(Current_Dir, "ua.locale"),
+            os.path.join(Target_Path, "ua.locale"),
+        )
+        console.print("Added Ukrainian locale", style="bold green")
+    except:
+        console.print("Failed to add Ukrainian locale", style="bold red")
